added athena utils file and removed repeated code in filter files
